Implement the fs enumerator for `backup`, don't defer in a loop (d'oh), cont'd. tidying
