+++ cmd/fleetctl/gitops-migrate/fs.go
@@ -8,6 +8,12 @@ import (
 	"path/filepath"
 )
 
+const (
+	fileModeUserRWX    = 0o700
+	fileModeUserRW     = 0o600
+	fileFlagsOverwrite = os.O_CREATE | os.O_WRONLY | os.O_TRUNC
+)
+
 // We use this error to break the 'fsEnum' iterator if, for whatever reason, the
 // caller breaks the iterator.
 var (
