+++ cmd/fleetctl/gitops-migrate/restore.go
@@ -1,0 +1,166 @@
+package main
+
+import (
+	"archive/tar"
+	"compress/gzip"
+	"context"
+	"errors"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+)
+
+// restore restores a Fleet GitOps backup from the provided tarball.
+//
+// 'from' must be a path to a gzipped tarball.
+//
+// If the 'to' path is not provided, the current working directory will be used.
+func restore(ctx context.Context, from string, to string) error {
+	log := LoggerFromContext(ctx)
+
+	// Set a default 'to' path, if necessary.
+	if to == "" {
+		log.Debug(
+			"found no 'to' path for restore operation, defaulting to " +
+				"current working directory",
+		)
+		to = "."
+	}
+	log.Info("performing Fleet GitOps restore", "from", from, "to", to)
+
+	// Create the output directory, if necessary.
+	err := os.MkdirAll(to, fileModeUserRWX)
+	if err != nil && !errors.Is(err, os.ErrExist) {
+		return fmt.Errorf(
+			"failed to create restore output directory: %w",
+			err,
+		)
+	}
+
+	// Get a readable handle to the archive.
+	//
+	//nolint:gosec,G304 // 'from' is a trusted input.
+	f, err := os.Open(from)
+	if err != nil {
+		return fmt.Errorf(
+			"failed to get a readable handle to restore archive(%s): %w",
+			from, err,
+		)
+	}
+
+	// Wrap the file stream in a gzip reader.
+	gz, err := gzip.NewReader(f)
+	if err != nil {
+		return fmt.Errorf(
+			"failed to create gzip reader from restore archive file stream(%s): %w",
+			from, err,
+		)
+	}
+
+	// Wrap the gzip reader in a tar reader.
+	tr := tar.NewReader(gz)
+
+	// Defer closure of all readers*, in reverse order.
+	//
+	// * Except the tar reader, it's not a 'ReadCloser'.
+	defer func() {
+		var errs error
+
+		// Close the gzip reader.
+		err := gz.Close()
+		if err != nil {
+			log.Error(
+				"failed to close restore archive gzip reader",
+				"error", err,
+			)
+		}
+		errs = errors.Join(errs, err)
+
+		// Close the restore archive file stream.
+		err = f.Close()
+		if err != nil {
+			log.Error(
+				"failed to close restore archive file stream",
+				"error", err,
+			)
+		}
+		errs = errors.Join(errs, err)
+
+		if errs != nil {
+			log.Error("errors encountered in restore archive stream close, exiting")
+			os.Exit(1)
+		}
+	}()
+
+	// Read and extract all files from the tarball.
+	for {
+		// Get the next compressed file.
+		header, err := tr.Next()
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				// We'll catch an 'io.EOF' when we hit the end of the archive.
+				return nil
+			}
+			// Otherwise, something has gone wrong.
+			return fmt.Errorf(
+				"failed to read next file from the restore archive stream: %w",
+				err,
+			)
+		}
+
+		// Construct the output path for this item.
+		output := filepath.Join(to, filepath.Clean(header.Name))
+		log.Info("decompressing restore archive item", "to", output)
+
+		// Handle the fs op based on the header type.
+		switch header.Typeflag {
+		case tar.TypeDir:
+			// Simply create the directory.
+			err := os.MkdirAll(output, fileModeUserRWX)
+			if err != nil && !errors.Is(err, os.ErrExist) {
+				return fmt.Errorf("failed to create output directory(%s): %w", output, err)
+			}
+
+		case tar.TypeReg:
+			// Get a writable handle to the restore target.
+			//
+			//nolint:gosec,G304 // 'output' is sanitized above.
+			decompressed, err := os.Create(output)
+			if err != nil {
+				return fmt.Errorf(
+					"failed to open writable stream to output file(%s): %w",
+					output, err,
+				)
+			}
+
+			// Decompress the file to disk.
+			//
+			//nolint:gosec,G110 // These are archives [hopefully] we created.
+			n, err := io.Copy(decompressed, tr)
+			if err != nil {
+				return fmt.Errorf(
+					"failed to decompress file(%s) during backup restoration: %w",
+					output, err,
+				)
+			}
+			// Make sure we wrote the expected content length.
+			if n != header.Size {
+				return fmt.Errorf(
+					"encountered no error in restore archive file(%s) decompression, "+
+						"but the archive's file size(%d) does not match what we wrote to disk(%d)",
+					output, header.Size, n,
+				)
+			}
+
+			// Close the output file.
+			err = decompressed.Close()
+			if err != nil {
+				return fmt.Errorf(
+					"failed to close output file(%s) stream during restore: %w",
+					output, err,
+				)
+			}
+		}
+	}
+}
